tidying up folders
put microcontroller and GUI into separate folder
