+++ GUI/GUI Python/Logo.png
